Fixed type warnings. Improved detection for points inside mesh: instead of relying on the normal of the closest primitive, we cast a ray on any direction from the position, and count the number of faces it crosses (odd: inside, even: outside)

+++ c++/src/TestMesh/helpers/glfw/input.h
@@ -291,8 +291,8 @@ namespace Mouse
         glfwGetCursorPos(window, &x, &y);
         mouse.dx = (f32)x - mouse.x;
         mouse.dy = (f32)y - mouse.y;
-        mouse.x = x;
-        mouse.y = y;
+        mouse.x = (f32)x;
+        mouse.y = (f32)y;
 	}
 };
